Remove PackageDb module.
It is superseeded by Sandbox
