+++ lib/PackageEnvironment.js
@@ -3,13 +3,10 @@
  */
 
 import type {
+  Sandbox,
+  PackageInfo,
   PackageJson,
   EnvironmentVarExport
-} from './PackageDb';
-
-import type {
-  Sandbox,
-  PackageInfo
 } from './Sandbox';
 
 const path = require('path');
