+++ lib/Sandbox.js
@@ -45,7 +45,33 @@ export type PackageInfo = {
   dependencyTree: DependencyTree;
 };
 
-export type PackageJson = any;
+export type PackageJsonVersionSpec = {
+  [name: string]: string;
+};
+
+export type EnvironmentVarExport = {
+  val: string;
+  scope?: string;
+  exclusive?: boolean;
+  __BUILT_IN_DO_NOT_USE_OR_YOU_WILL_BE_PIPd?: boolean;
+};
+
+export type PackageJson = {
+  name: string;
+  version?: string;
+  dependencies?: PackageJsonVersionSpec;
+  peerDependencies?: PackageJsonVersionSpec;
+  devDependencies?: PackageJsonVersionSpec;
+  optionalDependencies?: PackageJsonVersionSpec;
+
+  pjc?: {
+    build?: string;
+  };
+
+  exportedEnvVars?: {
+    [name: string]: EnvironmentVarExport;
+  }
+};
 
 export type DependencyTree = {
   [dependencyName: string]: PackageInfo;
